melhorias nas classes para teste com dados reais

+++ lib/packet/packet.cpp
@@ -48,7 +48,7 @@ void packet::safetyPacket(uint8_t ID, uint8_t deviceType, double latitude, doubl
     memcpy(returnPacket, &pkt, sizeof(SafetyPayload));
 }
 
-void packet::monitoringPacket(uint8_t ID, uint8_t deviceType, double latitude, double longitude, uint8_t batteryLevel, int32_t last5positions[5][2], uint8_t last5events[5], uint8_t status, uint8_t *returnPacket) {
+void packet::monitoringPacket(uint8_t ID, uint8_t deviceType, int32_t latitude, int32_t longitude, uint8_t batteryLevel, int32_t last5positions[5][2], uint8_t last5events[5], uint8_t status, uint8_t *returnPacket) {
     MonitoringPayload pkt;
 
     pkt.packetType = MONITORING_PACKET;
